rename utils to lib
